// src/shared/api/generated/models/CoworkingAccountBookingResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CRM API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CoworkingAccountBookingResponse
 */
export interface CoworkingAccountBookingResponse {
    /**
     * 
     * @type {string}
     * @memberof CoworkingAccountBookingResponse
     */
    login: string;
    /**
     * 
     * @type {Date}
     * @memberof CoworkingAccountBookingResponse
     */
    from: Date;
    /**
     * 
     * @type {Date}
     * @memberof CoworkingAccountBookingResponse
     */
    to: Date;
}

/**
 * Check if a given object implements the CoworkingAccountBookingResponse interface.
 */
export function instanceOfCoworkingAccountBookingResponse(value: object): value is CoworkingAccountBookingResponse {
    if (!('login' in value) || value['login'] === undefined) return false;
    if (!('from' in value) || value['from'] === undefined) return false;
    if (!('to' in value) || value['to'] === undefined) return false;
    return true;
}

export function CoworkingAccountBookingResponseFromJSON(json: any): CoworkingAccountBookingResponse {
    return CoworkingAccountBookingResponseFromJSONTyped(json, false);
}

export function CoworkingAccountBookingResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CoworkingAccountBookingResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'login': json['login'],
        'from': (new Date(json['from'])),
        'to': (new Date(json['to'])),
    };
}

export function CoworkingAccountBookingResponseToJSON(json: any): CoworkingAccountBookingResponse {
    return CoworkingAccountBookingResponseToJSONTyped(json, false);
}

export function CoworkingAccountBookingResponseToJSONTyped(value?: CoworkingAccountBookingResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'login': value['login'],
        'from': ((value['from']).toISOString()),
        'to': ((value['to']).toISOString()),
    };
}